+++ test/unit/orm.test.js
@@ -340,4 +340,71 @@ config.parallel('orm.test.js', () => {
             });
         });
     });
+    describe('ISSUES', () => {
+        it('#791 Document methods are not bind() to the document', async () => {
+            const db = await RxDB.create({
+                name: util.randomCouchString(),
+                adapter: 'memory',
+                multiInstance: false
+            });
+
+            const schema = {
+                version: 0,
+                type: 'object',
+                primaryPath: '_id',
+                properties: {
+                    name: {
+                        type: 'string'
+                    },
+                    nested: {
+                        type: 'object',
+                        properties: {
+                            foo: {
+                                type: 'string'
+                            }
+                        }
+                    }
+                }
+            };
+
+            const collection = await db.collection({
+                name: 'person',
+                schema: schema,
+                methods: {
+                    hello: function() {
+                        return this.name;
+                    }
+                }
+            });
+
+            const doc = await collection.insert({
+                name: 'hi',
+                nested: {
+                    foo: 'bar'
+                }
+            });
+
+            // orm-method
+            const hello = doc.hello;
+            assert.equal(hello(), 'hi');
+
+            // prototype-method
+            const get = doc.get;
+            assert.equal(get('name'), 'hi');
+
+            // nested
+            const nestedObj = doc.nested;
+            assert.equal(nestedObj.foo, 'bar');
+
+            // nested getter-method
+            const obs = nestedObj.foo$;
+            const emitted = [];
+            const sub = obs.subscribe(v => emitted.push(v));
+            await AsyncTestUtil.waitUntil(() => emitted.length === 1);
+            assert.equal(emitted[0], 'bar');
+            sub.unsubscribe();
+
+            db.destroy();
+        });
+    });
 });