+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Features:
   - Added `RxReplicationState.denied$` [#763](https://github.com/pubkey/rxdb/issues/763)
   - Added option for CORS to server-plugin
   - `this`-scope of collection-hooks are bound to the collection itself [#788](https://github.com/pubkey/rxdb/issues/788)
+  - All methods of `RxDocument` are bound to the instance [#791](https://github.com/pubkey/rxdb/issues/791)
 
 Bugfixes:
   - checkAdapter doesn't cleanup test databases [#714](https://github.com/pubkey/rxdb/issues/714)